restructure code a lot to match a defined architecture
