+++ src/app/layout/AppLayout.tsx
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import React, { FunctionComponent, PropsWithChildren, ReactNode } from 'react';
+
+export type AppLayoutProps = PropsWithChildren<{
+    topBar: ReactNode;
+}>;
+
+const AppLayout: FunctionComponent<AppLayoutProps> = ({ topBar, children }) => {
+    return (
+        <>
+            {topBar}
+            {children}
+        </>
+    );
+};
+
+export default AppLayout;
