+++ src/shared/lib/config-params.ts
