+++ src/shared/lib/error.ts
